+++ src/crud/players.py
@@ -6,7 +6,7 @@ from src.models.user import User
 from src.models.tournament import TournamentParticipants
 #from src.models.match import Match
 
-from src.schemas.player import CreatePlayerRequest #PlayerUpdate
+from src.schemas.player import CreatePlayerRequest, PlayerUpdate
 
 
 def create_player(db: Session, request: CreatePlayerRequest):
@@ -109,26 +109,34 @@ def read_all_players(db: Session, tournament_id: UUID | None = None):
 #     db.refresh(player)
 #     return player
 
-# def update_player(db: Session, player_id: uuid, updates: PlayerUpdate) -> Player:
-#     player = db.query(Player).filter(Player.id == player_id).first()
-#     if player:
-#         for key, value in updates.model_dump(exclude_unset=True).items():
-#             setattr(player, key, value)
-#         db.commit()
-#         db.refresh(player)
-#     else:
-#         raise HTTPException(
-#             status_code=status.HTTP_404_NOT_FOUND, detail="Player not found"
-#         )
-#     return player
+def update_player(db: Session, player_id: UUID, updates: PlayerUpdate, current_user: User) -> Player:
+    player = db.query(Player).filter(Player.id == player_id).first()
+    if not player:
+        raise HTTPException(
+                    status_code=status.HTTP_404_NOT_FOUND, detail="Player not found"
+                )
+    if player.user_id and (player.user_id != current_user.id or current_user.role != 'USER'):
+        raise HTTPException(
+                    status_code=status.HTTP_403_FORBIDDEN, detail="Current user can't update player"
+                )
+    for key, value in updates.model_dump(exclude_unset=True).items():
+        setattr(player, key, value)
+    db.commit()
+    db.refresh(player)
+
+    return player
 
 
-def delete_player(db: Session, player_id: UUID):
+def delete_player(db: Session, player_id: UUID, current_user: User):
     player = db.query(Player).filter_by(id=player_id).first()
     if not player:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Player not found"
         )
+    if (player.user_id and player.user_id != current_user.id) and current_user.role != 'ADMIN':
+        raise HTTPException(
+                    status_code=status.HTTP_403_FORBIDDEN, detail="Current user can't delete player"
+                )
     db.delete(player)
     db.commit()
     return True
@@ -140,7 +148,6 @@ def update_player_with_user(db: Session, player_id: UUID, user_id: UUID):
             status_code=status.HTTP_404_NOT_FOUND, detail="Player not found"
         )
 
-    # Check if the user exists
     user = db.query(User).filter_by(id=user_id).first()
     if not user:
         raise HTTPException(
@@ -152,3 +159,50 @@ def update_player_with_user(db: Session, player_id: UUID, user_id: UUID):
     db.commit()
     db.refresh(player)
     return player
+
+
+def read_player_by_id_and_tournament(db: Session, player_id: UUID, tournament_id: UUID):
+    player = (
+        db.query(
+            Player.id,
+            Player.first_name,
+            Player.last_name,
+            Player.country,
+            Player.team_id,
+            Player.matches_played,
+            Player.wins,
+            Player.losses,
+            Player.draws,
+            Player.user_id,
+            TournamentParticipants.tournament_id,
+            TournamentParticipants.score,
+            TournamentParticipants.stage
+        )
+        .join(TournamentParticipants, Player.id == TournamentParticipants.player_id)
+        .filter(Player.id == player_id, TournamentParticipants.tournament_id == tournament_id)
+        .first()
+    )
+
+    if not player:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, 
+            detail="Player not found in the specified tournament"
+        )
+    
+    player_data = {
+        "id": player.id,
+        "first_name": player.first_name,
+        "last_name": player.last_name,
+        "country": player.country,
+        "team_id": player.team_id,
+        "matches_played": player.matches_played,
+        "wins": player.wins,
+        "losses": player.losses,
+        "draws": player.draws,
+        "user_id": player.user_id,
+        "tournament_id": player.tournament_id,
+        "score": player.score,
+        "stage": player.stage
+    }
+    
+    return player_data
